+++ MyHomework_Lesson_1/MyHomework_Lesson_1.sln
@@ -3,9 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32002.261
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyHomework_Lesson_1_1", "MyHomework_Lesson_1_1\MyHomework_Lesson_1_1.csproj", "{D612FA5F-015D-42D3-B96A-582238F5372B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyHomework_Lesson_1_2", "MyHomework_Lesson_1_2\MyHomework_Lesson_1_2.csproj", "{A8318BEC-9023-4B25-BD0D-BE1B6FD9D2DA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyHomework_Lesson_1_1", "MyHomework_Lesson_1_1\MyHomework_Lesson_1_1.csproj", "{D612FA5F-015D-42D3-B96A-582238F5372B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +15,6 @@ Global
 		{D612FA5F-015D-42D3-B96A-582238F5372B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D612FA5F-015D-42D3-B96A-582238F5372B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D612FA5F-015D-42D3-B96A-582238F5372B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A8318BEC-9023-4B25-BD0D-BE1B6FD9D2DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8318BEC-9023-4B25-BD0D-BE1B6FD9D2DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8318BEC-9023-4B25-BD0D-BE1B6FD9D2DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8318BEC-9023-4B25-BD0D-BE1B6FD9D2DA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
